Add group CRU actions for displayname

+++ src/Moira.Authentik/Authentication/AuthentikAuthenticationResponseBody.cs
@@ -1,3 +1,3 @@
 namespace Moira.Authentik.Authentication;
 
-public record AuthentikAuthenticationResponseBody(string access_token, int expires_in, string token_type, string scope);
+public record AuthentikAuthenticationResponseBody(string access_token, long expires_in, string token_type, string scope);
